requirements.txt
crewai==0.141.0
python-dotenv
google-cloud-texttospeech
pydantic
langchain-google-genai==0.0.9 
google-generativeai
